+++ doc/Changelog.md
@@ -4,6 +4,7 @@
 
 - The PDF file converter accepts only one author field now.
 - Creates PDF/A-2b instead of PDF/A-1b
+- Add bookmarks (table of content) to PDFs
 
 ## v1.0 - 2019-03-01